LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/transition-finished-mark-as-handled-expected.txt


PASS navigation.transition.finished must not trigger unhandled rejections null is not an object (evaluating 'navigation.transition.finished')
